Simplify project config for `ProjectReference`
To enable use of the `SmtpServer` library as a `ProjectReference`
locally, this simplifies the ".csproj" file by removing the NuGet
packaging attributes and the license bundling. This does not remove
the root "LICENSE" file from the original fork and makes to attempt
to claim any ownership; this is merely to simplify the local use and
point to the correct repo of use.

This additionally upgrades the `System.IO.Pipelines` package `Version`
to "7.0.0", the latest stable publication at the time of editing.
However, this leaves the `LangVersion` at "8.0" since there is a use
of a now-impermissible buffer peeking behavior.